[git] Fix error fetching items from an empty repository
On empty repositories, those which do not have any history or
are only initialized, there are some commands that cannot
be run, like a pull or log. If any of this commands is called
an error will be returned. Thus, fetch() method fails when
a repository is empty.

This commit fixes this problem, generating an empty iterator
for those calls on an empty repository.

Backend version bumped to '0.6.1'.

Fixes #102

+++ perceval/backends/core/git.py
@@ -57,7 +57,7 @@ class Git(Backend):
     :raises RepositoryError: raised when there was an error cloning or
         updating the repository.
     """
-    version = '0.6.0'
+    version = '0.6.1'
 
     def __init__(self, uri, gitpath, tag=None, cache=None):
         origin = uri
@@ -97,6 +97,7 @@ class Git(Backend):
             branches_text = "no"
         else:
             branches_text = ", ".join(branches)
+
         logger.info("Fetching commits: '%s' git repository from %s; %s branches",
                     self.uri, str(from_date), branches_text)
 
@@ -110,9 +111,12 @@ class Git(Backend):
         ncommits = 0
         commits = self.__fetch_and_parse_log(from_date, branches)
 
-        for commit in commits:
-            yield commit
-            ncommits += 1
+        try:
+            for commit in commits:
+                yield commit
+                ncommits += 1
+        except EmptyRepositoryError:
+            pass
 
         logger.info("Fetch process completed: %s commits fetched",
                     ncommits)
@@ -130,7 +134,11 @@ class Git(Backend):
             repo = GitRepository.clone(self.uri, self.gitpath)
         elif os.path.isdir(self.gitpath):
             repo = GitRepository(self.uri, self.gitpath)
-        repo.pull()
+
+        try:
+            repo.pull()
+        except EmptyRepositoryError:
+            pass
 
         return repo
 
@@ -582,7 +590,6 @@ class EmptyRepositoryError(RepositoryError):
     message = "%(repository)s is empty"
 
 
-
 class GitRepository:
     """Manage a Git repository.
 
@@ -670,17 +677,24 @@ class GitRepository:
         'origin' repository. Any commit stored in the local copy will
         be removed.
 
+        :raises EmptyRepositoryError: when the repository is empty and
+            the action cannot be performed
         :raises RepositoryError: when an error occurs updating the
             repository
         """
+        if self.is_empty():
+            logger.warning("Git %s repository is empty; unable to pull",
+                           self.uri)
+            raise EmptyRepositoryError(repository=self.uri)
+
         cmd_fetch = ['git', 'fetch', 'origin']
         cmd_reset = ['git', 'reset', '--hard', 'origin']
 
         self._exec(cmd_fetch, cwd=self.dirpath, env={'LANG' : 'C'})
         self._exec(cmd_reset, cwd=self.dirpath, env={'LANG' : 'C'})
 
-        logging.debug("Git %s repository pulled into %s",
-                      self.uri, self.dirpath)
+        logger.debug("Git %s repository pulled into %s",
+                     self.uri, self.dirpath)
 
     def log(self, from_date=None, branches=None, encoding='utf-8'):
         """Read the commit log from the repository.
@@ -707,8 +721,14 @@ class GitRepository:
 
         :returns: a generator where each item is a line from the log
 
+        :raises EmptyRepositoryError: when the repository is empty and
+            the action cannot be performed
         :raises RepositoryError: when an error occurs fetching the log
         """
+        if self.is_empty():
+            logger.warning("Git %s repository is empty; unable to get the log",
+                           self.uri)
+            raise EmptyRepositoryError(repository=self.uri)
 
         cmd_log = ['git', 'log', '--raw', '--numstat', '--pretty=fuller',
                    '--decorate=full', '--reverse', '--topo-order',
